moved enums into server folder
