+++ server/frontend/package.json
@@ -47,9 +47,9 @@
     "typescript": "^4.0.2"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
+    "start": "react-app-rewired start",
+    "build": "react-app-rewired build",
+    "test": "react-app-rewired test",
     "eject": "react-scripts eject",
     "lint:fix": "eslint src/**/*.{js,jsx} --fix"
   },
@@ -70,6 +70,8 @@
   },
   "devDependencies": {
     "@types/moment": "^2.13.0",
-    "cross-env": "^7.0.2"
+    "cross-env": "^7.0.2",
+    "customize-cra": "^1.0.0",
+    "react-app-rewired": "^2.1.6"
   }
 }
