+++ server/frontend/src/types/Policy/AdaptionPolicy/ContentFlags/WordContentFlags.ts
@@ -1,4 +1,4 @@
-import { ContentManagementFlagAction } from "../../../../enums/ContentManagementFlagAction";
+import { ContentManagementFlagAction } from "../../../../../../src/common/models/enums/ContentManagementFlagAction";
 
 export interface WordContentFlags {
     dynamicDataExchange: ContentManagementFlagAction,
