+++ server/frontend/src/types/Policy/NcfsPolicy/NcfsOptions.ts
@@ -1,4 +1,4 @@
-import { NcfsOption } from "../../../enums/NcfsOption";
+import { NcfsOption } from "../../../../../src/common/models/enums/NcfsOption";
 
 export interface NcfsOptions {
     unProcessableFileTypes: NcfsOption,
