+++ server/src/common/models/TransactionEventService/GetTransactions/GetTransactions.test.ts
@@ -1,6 +1,6 @@
 import { Guid } from "guid-typescript";
-import { FileType } from "../../../../../frontend/src/enums/FileType";
-import { Risk } from "../../../../../frontend/src/enums/Risk";
+import { FileType } from "../../enums/FileType";
+import { Risk } from "../../enums/Risk";
 import { GetTransactionsRequest, GetTransactionsResponse } from "./index";
 
 describe("GetTransactions", () => {
