+++ server/src/common/models/enums/FileDetailsStatus.ts
