+++ server/src/common/models/enums/FileType.ts
