+++ server/src/common/models/enums/NcfsOption.ts
