+++ server/src/common/models/enums/PolicyType.ts
