+++ server/src/service/Routes/RequestHistory/RequestHistoryRoutes.test.ts
@@ -3,8 +3,8 @@ import bodyParser from "body-parser";
 import winston from "winston";
 import { SinonStub, stub } from "sinon";
 import request from "supertest";
-import { Risk } from "../../../../frontend/src/enums/Risk";
-import { FileType } from "../../../../frontend/src/enums/FileType";
+import { Risk } from "../../../common/models/enums/Risk";
+import { FileType } from "../../../common/models/enums/FileType";
 import { GetTransactionsResponse } from "../../../common/models/TransactionEventService/GetTransactions";
 import { GetTransactionDetailsResponse } from "../../../common/models/TransactionEventService/GetTransactionDetails";
 import TestConfig from "../../TestConfig";
